requirements_only_run.txt
PyPDF2==1.26.0
Pillow==8.1.0